Tap Mechanic tiens Adrien kiss kiss

+++ MobileProject2021/Assets/Scripts/Managers/EnemyManager/OnSpawnEventArgs.cs
@@ -8,5 +8,5 @@ public class OnSpawnEventArgs : EventArgs
     public SO_Enemy soEnemy;
     public SO_Boss soBoss;
     public float enemyLevel;
-    public int bossTimer;
+    public float bossTimer;
 }
